refine "replicas" into "collections"
a replica is one copy of a file somewhere. the replica is housed in a
"collection", which is the term for some pile of data that may have some
shared purpose

for example, for files to be replicated to a hof daemon, the daemon must
delcare a default collection for incoming files to be placed at

likewise, collections provide a slightly more natural way to fetch
files, through a new /collection/<collection name>/<file path> route.

another kind of collection that is not well implemented or documented
yet is "hams", for "hash-addressed managed storage". this is intended to
support the original idea of file indexing, where the exact properties
of data storage are left to whatever is appropriate for the data in
question. a common expected use of hams collections would be, for
example, to zstd compress .fits files to reduce at-rest space
consumption. none of this is implemented yet, of course.

as a general rule, hams-backed files are accessible as
<collection_name>:<filetag hams:path>, which automatically resolves to
the correct file with any storage transformation unapplied at fetch
time.